refactors to minimize line number in ObjectConverter
